manages to solve easy problems now!

+++ src/main/java/no/sintef/vrp/vrp_backend/api/vrp/schema.java
@@ -90,16 +90,16 @@ public class schema {
             this.name = name;
         }
 
-        public List<Long> getDistances() {
+        public List<Integer> getDistances() {
             return distances;
         }
 
-        public void setDistances(List<Long> distances) {
+        public void setDistances(List<Integer> distances) {
             this.distances = distances;
         }
 
         String name;
-        List<Long> distances;
+        List<Integer> distances;
     }
     public static class NewProblem {
         @Override
